pyproject.toml
[tool.poetry]
name = "ipl-config"
version = "0.1.0"
description = "InPlat config adapters"
authors = ["InPlat", "Valentin Nagornyy <valentin.nagornyy@inplat.ru>"]
repository = ""

[[tool.poetry.source]]
name = "inplat"
url = "http://pip.sys.ipl/simple"
# default = true

[tool.poetry.dependencies]
python = "^3.10"
ipapp = { version = "^1.3.7", extras = [ "fastapi", "postgres" ] }
uvloop = "^0.16.0"
python-dotenv = {version = "^0.19.2", optional = true}
# TODO: pyyaml = {version = "^6.0", optional = true}
pyyaml = {version = "^5.4", optional = true}
toml = {version = "^0.10", optional = true}

[tool.poetry.extras]
dotenv = ["python-dotenv"]
yaml = ["pyyaml"]
toml = ["toml"]

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.18.1"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"
flake8-comprehensions = "^3.8.0"
flake8-bugbear = "^22.1.11"
flake8-mutable = "^1.2.0"
pep8-naming = "^0.12.1"
pylint = "^2.12.2"
bandit = "^1.7.2"
black = "^22.1.0"
coverage = "^6.0.0"
safety = "^2.1.1"
mypy = "^0.941"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]

[tool.black]
line-length = 79
skip-string-normalization=true

[tool.isort]
multi_line_output = 3
line_length = 79
include_trailing_comma=true
lines_after_imports = 2

[tool.bandit]
skips = "B104"  # hardcoded_bind_all_interfaces
